requirements.txt
Django==1.10.3

django-filter==1.0.0

django-widget-tweaks==1.4.1

